created server, bug fixes
