--- models/reaction.ts
@@ -1,34 +1,0 @@
-import { Schema, Types, Document, ObjectId } from 'mongoose';
-import parseDate from '../utils/parseDate.js';
-
-interface IReaction extends Document {
-    reactionId: ObjectId;
-    reactionBody: string;
-    username: string;
-    createdAt: Date;
-};
-
-const reactionSchema = new Schema<IReaction>(
-    {
-        reactionId: {
-            type: Schema.Types.ObjectId,
-            default: () => new Types.ObjectId()
-        },
-        reactionBody: {
-            type: String,
-            required: true,
-            maxlength: 280
-        },
-        username: {
-            type: String,
-            required: true
-        },
-        createdAt:{
-            type: Date,
-            default: Date.now,
-            get: parseDate
-        }
-    }
-);
-
-export default reactionSchema;
