--- routes/api/friendRoutes.ts
@@ -1,12 +1,0 @@
-import { Router } from "express";
-
-const router = Router();
-
-import { addFriend, betrayal } from "../../controlers/friendController.js";
-
-router
-    .route('/:friendId')
-    .post(addFriend)
-    .delete(betrayal);
-
-export default router
