--- routes/api/thoughtIdRoutes.ts
@@ -1,17 +1,0 @@
-import { Router } from "express";
-
-const router = Router();
-
-import reactionRoutes from "./reactionRoutes"
-
-import { getOneThought, updateThought, headEmpty } from "../../controlers/thoughtController";
-
-router.use('/reactions', reactionRoutes)
-
-router
-    .route('/')
-    .get(getOneThought)
-    .put(updateThought)
-    .delete(headEmpty);
-
-export default router
