+++ src/config/connection.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
 
-mongoose.connect('mongodb://127.0.0.1:27017/developersApplications');
+mongoose.connect('mongodb://127.0.0.1:27017/SocialAPI');
 
 export default mongoose.connection;
