+++ src/utils/data.ts
@@ -1,0 +1,73 @@
+// const userNames = [
+//     'Aaran',
+//     'Aaren',
+//     'Aarez',
+//     'Aarman',
+//     'Aaron',
+//     'Aaron-James',
+//     'Aarron',
+//     'Aaryan',
+//     'Aaryn',
+//     'Aayan',
+//     'Aazaan',
+//     'Abaan',
+//     'Abbas',
+//     'Abdallah',
+//     'Abdalroof',
+//     'Abdihakim',
+//     'Abdirahman',
+//     'Abdisalam',
+//     'Abdul',
+//     'Abdul-Aziz',
+//     'Abdulbasir',
+//     'Abdulkadir',
+//     'Abdulkarem',
+//     'Smith',
+//     'Jones',
+//     'Coollastname',
+//     'enter_name_here',
+//     'Ze',
+//     'Zechariah',
+//     'Zeek',
+//     'Zeeshan',
+//     'Zeid',
+//     'Zein',
+//     'Zen',
+//     'Zendel',
+//     'Zenith',
+//     'Zennon',
+//     'Zeph',
+//     'Zerah',
+//     'Zhen',
+//     'Zhi',
+//     'Zhong',
+//     'Zhuo',
+//     'Zi',
+//     'Zidane',
+//     'Zijie',
+//     'Zinedine',
+//     'Zion',
+//     'Zishan',
+//     'Ziya',
+//     'Ziyaan',
+//     'Zohaib',
+//     'Zohair',
+//     'Zoubaeir',
+//     'Zubair',
+//     'Zubayr',
+//     'Zuriel',
+//     'Xander',
+//     'Jared',
+//     'Grace',
+//     'Alex',
+//     'Mark',
+//     'Tamar',
+//     'Farish',
+//     'Sarah',
+//     'Nathaniel',
+//     'Parker',
+//   ];
+
+//   const email = [
+//     'email@email.com'
+//   ]
