+++ src/utils/parseDate.ts
@@ -1,0 +1,3 @@
+const parseDate = (date: Date)  => date.toDateString();
+
+export default parseDate;
