--- utils/parseDate.ts
@@ -1,3 +1,0 @@
-const parseDate = (date) => date.toDateString();
-
-export default parseDate;
